+++ src/core/modules/atomic/warroom/__init__.py
@@ -5,6 +5,7 @@
 from .discover import WarroomDiscoverModule
 from .generate_scenarios import WarroomGenerateScenariosModule
 from .llm_review import WarroomLlmReviewModule
+from .public_site import WarroomPublicSiteVerifyModule
 from .report import WarroomReportModule
 from .run import WarroomRunModule
 
@@ -12,6 +13,7 @@ __all__ = [
     "WarroomDiscoverModule",
     "WarroomGenerateScenariosModule",
     "WarroomLlmReviewModule",
+    "WarroomPublicSiteVerifyModule",
     "WarroomReportModule",
     "WarroomRunModule",
 ]